# Optimize Concurrent::Map#[] on CRuby by letting the backing Hash handle the default_proc