--- a/lib/concurrent-ruby/concurrent/collection/map/mri_map_backend.rb
+++ b/lib/concurrent-ruby/concurrent/collection/map/mri_map_backend.rb
@@ -9,8 +9,8 @@ module Concurrent
     # @!visibility private
     class MriMapBackend < NonConcurrentMapBackend
 
-      def initialize(options = nil)
-        super(options)
+      def initialize(options = nil, &default_proc)
+        super(options, &default_proc)
         @write_lock = Mutex.new
       end