* On ruby 3.1.3p185 (2022-11-24 revision 1a6b16756e) [x86_64-linux]
      Before:
      Hash#[]     14.477M (± 2.7%) i/s -     72.882M in   5.038078s
       Map#[]      7.837M (± 1.7%) i/s -     39.947M in   5.098411s
      After:
      Hash#[]     14.340M (± 1.5%) i/s -     72.074M in   5.027414s
       Map#[]      9.840M (± 0.8%) i/s -     50.106M in   5.092390s

With this, `Concurrent::Map#[]` on CRuby is just:
```ruby
      def [](key)
        @backend[key] # @backend is the Hash
      end
```
i.e., just one extra call and ivar read compared to `Hash#[]`.