@eregon 2023-01-27
I also measured on `ruby 3.2.0 (2022-12-25 revision a528908271) [x86_64-linux]` and with YJIT.
```
Before:
             Hash#[]     16.102M (± 1.9%) i/s -     81.399M in   5.057130s
              Map#[]      8.296M (± 1.0%) i/s -     41.992M in   5.061993s
With YJIT:
             Hash#[]     17.601M (± 0.0%) i/s -     88.445M in   5.025111s
              Map#[]     17.058M (± 0.3%) i/s -     85.718M in   5.025048s

After:
             Hash#[]     15.736M (± 4.0%) i/s -     79.505M in   5.061376s
              Map#[]     11.846M (± 1.4%) i/s -     59.857M in   5.053804s
With YJIT:
             Hash#[]     18.469M (± 1.6%) i/s -     92.881M in   5.030200s
              Map#[]     15.542M (± 0.0%) i/s -     77.977M in   5.017096s
```

So YJIT reduces the difference between the two, cc @joeldrapper, and it looks already fast enough without changing anything with YJIT.
Surprisingly in that run the numbers look faster before this PR, might be noise or some issue in YJIT, unsure.