@eregon 2023-01-27
Some more measurements, 3.2 with YJIT:
```
Before:
Warming up --------------------------------------
             Hash#[]     1.646M i/100ms
              Map#[]     1.453M i/100ms
Calculating -------------------------------------
             Hash#[]     17.275M (± 1.0%) i/s -     87.212M in   5.048775s
              Map#[]     16.831M (± 1.2%) i/s -     84.251M in   5.006328s

Warming up --------------------------------------
             Hash#[]     1.647M i/100ms
              Map#[]     1.482M i/100ms
Calculating -------------------------------------
             Hash#[]     17.359M (± 0.6%) i/s -     87.286M in   5.028526s
              Map#[]     16.860M (± 1.0%) i/s -     84.501M in   5.012538s
```

```
After:
Warming up --------------------------------------
             Hash#[]     1.674M i/100ms
              Map#[]     1.621M i/100ms
Calculating -------------------------------------
             Hash#[]     19.580M (± 0.6%) i/s -     98.745M in   5.043264s
              Map#[]     16.564M (± 0.3%) i/s -     84.267M in   5.087353s

Warming up --------------------------------------
             Hash#[]     1.649M i/100ms
              Map#[]     1.606M i/100ms
Calculating -------------------------------------
             Hash#[]     19.545M (± 1.3%) i/s -     98.949M in   5.063710s
              Map#[]     16.527M (± 1.6%) i/s -     83.508M in   5.054362s
```
Interestingly the numbers in `Warming up` match what I expected.
But the numbers `Calculating` nope, they are weird.